requirements.txt
asyncio==3.4.3
Markdown==3.3.4
motor==2.4.0
PyJWT==2.1.0
pymongo==3.11.4
PyYAML==5.4.1
redis==3.5.3
tornado==6.1
ua-parser==0.10.0
user-agents==2.2.0
